Add an all page for governance requests table.
